requirements.txt
PyQt6
pyaudio
vosk
ollama
pyttsx3
playsound==1.2.2
#simpleaudio
pycairo
PyGObject
SpeechRecognition
comtypes
psutil
gpustat
#pybind11
#setuptools
#piper-tts==1.2.0
coqui-tts
#mycroft-mimic3-tts[all]
silero
sounddevice